Reorg PMEM FreeList: global free entry pool + thread local entry cache
