+++ CMakeLists.txt
@@ -28,9 +28,10 @@ set(SOURCES
         engine/write_batch.cpp
         engine/hash_list.cpp
         engine/dram_allocator.cpp
-        engine/pmem_allocator.cpp
+        engine/pmem_allocator/pmem_allocator.cpp
         engine/thread_manager.cpp
-        engine/utils.cpp)
+        engine/utils.cpp
+        engine/pmem_allocator/free_list.cpp)
 
 
 # .a library
